Factor out Selection.
