--- docs/core/selections.rst
@@ -1,5 +1,0 @@
-Selections
-**********
-
-.. autoclass:: mira.core.Selection
-    :members:
