+++ mira/core/__init__.py
@@ -1,6 +1,5 @@
 from .annotation import Annotation, AnnotationConfiguration
 from .scene import Scene, SceneCollection
-from .selection import Selection
 from . import augmentations
 from . import callbacks
 from . import utils
