+++ mira/detectors/fasterrcnn.py
@@ -79,7 +79,10 @@ class FasterRCNN(Detector):
             [
                 mc.Annotation(
                     category=self.annotation_config[int(labelIdx) - 1],
-                    selection=mc.Selection(x1=x1, y1=y1, x2=x2, y2=y2),
+                    x1=x1,
+                    y1=y1,
+                    x2=x2,
+                    y2=y2,
                     score=score,
                 )
                 for (x1, y1, x2, y2), labelIdx, score in zip(
